A* pathFinding & Camera Movement Refined
Couple of progress updates. I have working animations for the player character along with a tile sheet for the main hub level. I also have created A* pathfinding that couples nicely with the Generic grid system that I implemented. Still need to figure out a way to save layouts for what is walkable and what isn't for the characters this system is given to. This will be the next step is to implement a movement System for NPCS.