worked more on inventory UI functionality.
Right now working on the player UI functionality with the knowledge that I will break the class apart more to make a more Generic InventoryUI Class. Just using the PlayerUI class as testing grounds. I have the Inventory displaying the correct amount of slots available. Along with a new scriptable object type called GameEvent which uses the observer pattern to have multiple subscribers listen to it to see if the data has changed. Currently that seems to not be working so it will be the next thing I will work on.